+++ go.mod
@@ -1,3 +1,3 @@
 module "github.com/maragudk/gomponents"
 
-go 1.15
+go 1.18
